protobuf

+++ shared-dependencies/transfer.go
@@ -8,11 +8,11 @@ type Transfer struct {
 	Id     uuid.UUID `json:"id"`
 	From   string    `json:"from"`
 	To     string    `json:"to"`
-	Amount int       `json:"amount"`
+	Amount float64   `json:"amount"`
 	Status string    `json:"status"`
 }
 
 type Balance struct {
-	Username string `json:"username"`
-	Balance  int    `json:"balance"`
+	Username string  `json:"username"`
+	Balance  float64 `json:"balance"`
 }
